requirements.txt
google-generativeai
python-dotenv==1.0.0
Flask==2.3.3
requests==2.31.0
urllib3==1.26.15
python-jenkins==1.8.0
flask-cors==4.0.0
gunicorn==21.2.0
beautifulsoup4==4.12.3
xmltodict==0.13.0
GitPython==3.1.43
